+++ tests/matchers/test_possible_call_matcher.py
@@ -164,3 +164,56 @@ def test_signature_contains_only_known_named_arguments_if_call_dont_contain_2sta
     assert PossibleCallMatcher('a, c').match(example)
     assert PossibleCallMatcher('b, c').match(example)
     assert PossibleCallMatcher().match(example)
+
+
+def test_random_functions(transformed):
+    @transformed
+    def function_1():
+        pass
+    @transformed
+    def function_2(arg):
+        pass
+    @transformed
+    def function_3(**kwargs):
+        pass
+    @transformed
+    def function_4(*args, **kwargs):
+        pass
+    @transformed
+    def function_5(a, b):
+        pass
+    @transformed
+    def function_6(a, b, c):
+        pass
+    @transformed
+    def function_7(a, b, c=False):
+        pass
+    @transformed
+    def function_8(a, b, c=False, *d):
+        pass
+    @transformed
+    def function_9(a, b, c=False, *d, **e):
+        pass
+    @transformed
+    def function_10(a, b, c=False, c2=False, *d, **e):
+        pass
+    @transformed
+    def function_11(a, b, b2, c=False, c2=False, *d, **e):
+        pass
+    @transformed
+    def function_12(c=False, c2=False):
+        pass
+
+    assert PossibleCallMatcher('.', '**').match(function_4)
+    assert PossibleCallMatcher('.', '.').match(function_7)
+    assert PossibleCallMatcher('.', '.', 'c').match(function_8)
+    assert PossibleCallMatcher('.', '.', 'c2', '*', '**').match(function_10)
+    assert PossibleCallMatcher('c').match(function_12)
+
+    assert not PossibleCallMatcher('.').match(function_1)
+    assert not PossibleCallMatcher('c').match(function_2)
+    assert not PossibleCallMatcher('.', '**').match(function_3)
+    assert not PossibleCallMatcher('.', 'c').match(function_5)
+    assert not PossibleCallMatcher('.', '.').match(function_6)
+    assert not PossibleCallMatcher('.', 'c', '*', '**').match(function_9)
+    assert not PossibleCallMatcher('.', '.', 'c2', '*', '**').match(function_11)
